+++ copy_dir.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strings"
 )
 
@@ -89,62 +88,4 @@ func copyDir(ctx context.Context, dst string, src string, ignoreDot bool, disabl
 	}
 
 	return filepath.Walk(resolved, walkFn)
-}
-
-// resolveSymlinks resolves symlinks with special handling for Windows junction points in Go 1.23+.
-// This function provides a more robust fallback for Windows path resolution issues.
-func resolveSymlinks(src string) (string, error) {
-	resolved, err := filepath.EvalSymlinks(src)
-	if err == nil {
-		return resolved, nil
-	}
-
-	// On non-Windows platforms, EvalSymlinks should work properly
-	if runtime.GOOS != "windows" {
-		return "", err
-	}
-
-	// On Windows, first check if the path actually exists and what type it is
-	lstatInfo, lstatErr := os.Lstat(src)
-	if lstatErr != nil {
-		// Path doesn't exist or can't be accessed, return the original EvalSymlinks error
-		return "", err
-	}
-
-	statInfo, statErr := os.Stat(src)
-	if statErr != nil {
-		// Path can't be followed, return the original EvalSymlinks error
-		return "", err
-	}
-
-	// Path exists but EvalSymlinks failed on Windows
-	// Check what type of path this is
-
-	// Case 1: Check if this is a junction point
-	var isJunction bool
-	var junctionErr error
-	if isJunction, junctionErr = isWindowsJunctionPoint(src); junctionErr == nil && isJunction {
-		// Confirmed junction point - return cleaned path
-		return filepath.Clean(src), nil
-	}
-
-	// Case 2: Check if this is a regular directory or file
-	isRegularDir := lstatInfo.IsDir() && statInfo.IsDir() &&
-		lstatInfo.Mode()&os.ModeIrregular == 0 &&
-		lstatInfo.Mode()&os.ModeSymlink == 0
-
-	if isRegularDir {
-		// I'm just a dir, yes I'm only a dir,
-		// EvalSymlinks tried but it stopped right here,
-		// So do what EvalSymlinks would normally do:
-		// Return the cleaned path, plain and true.
-		return filepath.Clean(src), nil
-	}
-
-	// If we get here, it's some other type of special file/link that we don't understand
-	// Return the original EvalSymlinks error
-	if junctionErr != nil {
-		return "", fmt.Errorf("failed to resolve symlinks (checking whether junction point: %s): %w", junctionErr, err)
-	}
-	return "", fmt.Errorf("failed to resolve symlinks: %w", err)
 }
